Updated to v2.0.0!
So, as I promised, we got a major update just after a month. 1M and 12D later to be exact. AND THIS BRINGS A LOT. I did so much of backend that now code is WAYY better than what it was. SO, WE GOT LEVELLING SYSTEM AND RANKINGS TOO! I changed the Bot Names, added MORE Countries! Made a whole new system just for this special project. It took a lot. It was in work since December 29, I guess? After solo testing, a lot of time spent, I finally present you the new world of OddEven. All the games in the market wouldn't be that good as mine, that's my guarantee. Now that I'm running out of ideas, its your turn now.